+++ docs/assets/chevron-down-0r6b9kiL.js
@@ -1,4 +1,4 @@
-import{c as o}from"./index-jgVsv-40.js";/**
+import{c as o}from"./index-CZL6KbO1.js";/**
  * @license lucide-react v1.31.0 - ISC
  *
  * This source code is licensed under the ISC license.