+++ docs/assets/user-RgqO82o1.js
@@ -1,4 +1,4 @@
-import{c}from"./index-jgVsv-40.js";/**
+import{c}from"./index-CZL6KbO1.js";/**
  * @license lucide-react v1.31.0 - ISC
  *
  * This source code is licensed under the ISC license.